+++ src/application/use-cases/delete-income.ts
@@ -5,9 +5,15 @@ export interface DeleteIncomeUseCaseRequest {
 }
 
 export class DeleteIncomeUseCase {
-  constructor(private incomesRepository: IncomesRepository) {}
+  constructor(private incomesRepository: IncomesRepository) { }
 
   async execute({ id }: DeleteIncomeUseCaseRequest): Promise<void> {
+    const existingIncome = await this.incomesRepository.findById(id)
+
+    if (!existingIncome) {
+      throw new Error('Income not found')
+    }
+
     await this.incomesRepository.delete(id)
   }
 }
